Add basic e-mail support.
Users can now trigger mail to be sent to reviewees,
which at this point is little more than a basic indicator
that someone has made comments about their review.

Improvements need to be made, including better error
handling for e-mail, and some measure of testing.

Additionally, a better e-mail message from the server
is also necessary, with the details of the review annotations
inlined.